pyproject.toml
[project]
name = "cicd-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
]
